requirements.txt
Flask==2.2.3
Werkzeug==2.2.3
Flask-RESTful==0.3.9
paho-mqtt==1.6.1
python-dotenv==0.21.1
flask-bootstrap4==4.0.2
uuid==1.30
Flask-Login==0.6.2
